requirements.txt
python-telegram-bot>=21.0
apscheduler>=3.10.0,<4.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
langfuse>=2.0.0
psutil>=5.9.0
# Security scanning — used by Sentinel
pip-audit>=2.7.0
bandit[toml]>=1.7.0

# Anthropic is optional — only needed if LLM_PROVIDER=anthropic
# anthropic>=0.40.0

# Zuko — job scraping and browser automation
requests>=2.32.0
playwright>=1.44.0
